Refactored entities and pin project, sort, search

+++ backend/src/main/java/com/szte/saturn/controllers/TicketController.java
@@ -33,7 +33,7 @@ public class TicketController {
     }
 
     @GetMapping("/")
-    public ResponseEntity<List<Ticket>> getAllByProjectsTickets(@PathVariable Integer projectId ) {
+    public ResponseEntity<List<Ticket>> getAllByProjectsTickets(@PathVariable Long projectId ) {
         List<Ticket> tickets = ticketService.getAllTickets(projectId);
 
         return ResponseEntity.ok(tickets);
