+++ backend/src/main/java/com/szte/saturn/repositories/UserRepository.java
@@ -7,10 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
-
-    Optional<User> findUserById(int id);
-
-    //Optional<User> findUserByProjectId(int id);
+    Optional<User> findUserById(Long id);
 }
